deleted unn. includes, fixed strsplit

+++ ft_strfnd.c
@@ -6,7 +6,7 @@
 /*   By: oevtushe <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/19 14:14:51 by oevtushe          #+#    #+#             */
-/*   Updated: 2017/11/19 14:21:15 by oevtushe         ###   ########.fr       */
+/*   Updated: 2017/11/20 08:07:23 by oevtushe         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int		ft_strfnd(const char *s, char c)
 {
-	int	i;
+	size_t	i;
 
 	i = 0;
 	while (s[i] != '\0')
